grpc_polyglot_demo/client_python/requirements.txt

grpcio
grpcio-tools
numpy
matplotlib
